rm decimal-places
