+++ services/payments/src/simcore_service_payments/api/rpc/_payments.py
@@ -50,9 +50,9 @@ async def init_payment(
         init = await payments_gateway_api.init_payment(
             payment=InitPayment(
                 amount_dollars=amount_dollars,
-                credits_=target_credits,
+                credits=target_credits,
                 user_name=user_name,
-                user_email=user_email,  # type: ignore
+                user_email=user_email,
                 wallet_name=wallet_name,
             )
         )
